v0.4 foundation6 menu template and admin:menus css l18n for sections first sections tests ajax for mail view, section template names search pattern replacements make submit translatable subscribe: add settings and renaming email in subscriber admin list logging and sortable enabled pagenode validation search_replacements.json in MANIFEST autopopulatedentry and sections.simple... menu mixin finally in menus rm pageprefix

+++ pagetools/subscribe/utils.py
@@ -41,7 +41,7 @@ def send_queued_mail(qm, maxmails):
     sts = SendStatus.objects.filter(queued_email=qm)[:maxmails]
     qm.send_to_all(sts)
     sst2 = SendStatus.objects.filter(queued_email=qm)
-    if not sst2:
+    if not sst2 and subs_settings.DELETE_QUEUED_MAILS:
         qm.delete()
     return len(sts)
